+++ src/components/MultiSelect/MultiSelectMenuButton.tsx
@@ -24,6 +24,14 @@ const grow = keyframes`
   from {width: 22px; opacity: 0; }
   to {width: 46px; opacity: 1;}
 `;
+
+/**
+ * The toggle button component used to open and close the `MultiSelect` menu.
+ * A second button is rendered above the main button that displays the current
+ * number of selected items. Clicking on the second button will clear all
+ * the selected items and the main button's close event will not be fired
+ * (as expected).
+ */
 const MultiSelectMenuButton = forwardRef<
   HTMLButtonElement,
   MultiSelectMenuButtonProps
@@ -39,15 +47,21 @@ const MultiSelectMenuButton = forwardRef<
     selectedItems,
     ...rest
   } = props;
-  const styles = useMultiStyleConfig("MultiSelectMenuButton", { isOpen });
   const iconType = isOpen ? "minus" : "plus";
   const growAnimation = `${grow} 150ms ease-out`;
 
   // Sets the selected items count on the menu button.
   let getSelectedItemsCount;
+  let selectedItemsAriaLabel;
   if (selectedItems[multiSelectId]?.items.length > 0) {
     getSelectedItemsCount = `${selectedItems[multiSelectId].items.length}`;
+    const itemPlural = getSelectedItemsCount === "1" ? "" : "s";
+    selectedItemsAriaLabel = `${getSelectedItemsCount} item${itemPlural} selected`;
   }
+  const styles = useMultiStyleConfig("MultiSelectMenuButton", {
+    isOpen,
+    hasSelectedItems: getSelectedItemsCount,
+  });
   // We need this for our "fake" button inside the main menu button.
   function onKeyPress(e) {
     const enterOrSpace =
@@ -64,17 +78,24 @@ const MultiSelectMenuButton = forwardRef<
   }
 
   return (
-    <Button
-      buttonType="secondary"
-      id={id}
-      onClick={onMenuToggle}
-      ref={ref}
-      __css={styles.menuButton}
-      {...rest}
-    >
+    <>
+      <Button
+        buttonType="secondary"
+        id={id}
+        onClick={onMenuToggle}
+        ref={ref}
+        __css={styles.menuButton}
+        {...rest}
+      >
+        <Box as="span" title={multiSelectLabel} __css={styles.buttonLabel}>
+          {multiSelectLabel}
+        </Box>
+        <Icon id={`ms-${multiSelectId}-icon`} name={iconType} size="small" />
+      </Button>
       {getSelectedItemsCount && (
         <Box
           animation={growAnimation}
+          aria-label={selectedItemsAriaLabel}
           as="span"
           onClick={onClear}
           onKeyPress={onKeyPress}
@@ -94,16 +115,7 @@ const MultiSelectMenuButton = forwardRef<
           />
         </Box>
       )}
-      <Box as="span" title={multiSelectLabel} __css={styles.buttonLabel}>
-        {multiSelectLabel}
-      </Box>
-      <Icon
-        id={`ms-${multiSelectId}-icon`}
-        name={iconType}
-        size="small"
-        __css={styles.toggleIcon}
-      />
-    </Button>
+    </>
   );
 });