+++ src/hooks/__tests__/useMultiSelect.test.ts
@@ -20,66 +20,62 @@ const items = [
 
 describe("useMultiSelect hook", () => {
   it("should return three functions and a object containing a key items with all currently selectItems", () => {
-    const { result } = renderHook(() =>
-      useMultiSelect({ multiSelectId: multiSelectId, items })
-    );
+    const { result } = renderHook(() => useMultiSelect());
     expect(typeof result.current.onChange).toEqual("function");
     expect(typeof result.current.onMixedStateChange).toEqual("function");
     expect(typeof result.current.onClear).toEqual("function");
     expect(result.current.selectedItems).toEqual({});
   });
   it("should update the selectedItems when the onChange function is called", () => {
-    const { result } = renderHook(() =>
-      useMultiSelect({ multiSelectId: multiSelectId, items })
-    );
+    const { result } = renderHook(() => useMultiSelect());
     // selectedItems starts empty
     expect(result.current.selectedItems).toEqual({});
 
-    act(() => result.current.onChange(items[0].id));
+    act(() => result.current.onChange(items[0].id, multiSelectId));
     // Add correct id to items array
     expect(result.current.selectedItems).toEqual({
       "multiselect-hook-test-id": { items: ["dogs"] },
     });
 
-    act(() => result.current.onChange(items[1].id));
+    act(() => result.current.onChange(items[1].id, multiSelectId));
     // Adds second item to the selectedItems
     expect(result.current.selectedItems).toEqual({
       "multiselect-hook-test-id": { items: ["dogs", "cats"] },
     });
   });
   it("should update the selectedItems when the onMixedstateChange function is called", () => {
-    const { result } = renderHook(() =>
-      useMultiSelect({ multiSelectId: multiSelectId, items })
-    );
+    const { result } = renderHook(() => useMultiSelect());
 
     // selectedItems starts empty
     expect(result.current.selectedItems).toEqual({});
 
-    act(() => result.current.onMixedStateChange(items[3].id));
+    act(() =>
+      result.current.onMixedStateChange(items[3].id, multiSelectId, items)
+    );
     // Both child items should be in the selectedItems
     expect(result.current.selectedItems).toEqual({
       "multiselect-hook-test-id": { items: ["red", "blue"] },
     });
   });
   it("should return again a empty object of selectedItems when the onClear function is called", () => {
-    const { result } = renderHook(() =>
-      useMultiSelect({ multiSelectId: multiSelectId, items })
-    );
+    const { result } = renderHook(() => useMultiSelect());
     // selectedItems starts empty
     expect(result.current.selectedItems).toEqual({});
 
-    act(() => result.current.onMixedStateChange(items[3].id));
+    act(() =>
+      result.current.onMixedStateChange(items[3].id, multiSelectId, items)
+    );
     // Both child items should be in selectedItems
     expect(result.current.selectedItems).toEqual({
       "multiselect-hook-test-id": { items: ["red", "blue"] },
     });
 
-    act(() => result.current.onChange(items[0].id));
+    act(() => result.current.onChange(items[0].id, multiSelectId));
     expect(result.current.selectedItems).toEqual({
       "multiselect-hook-test-id": { items: ["red", "blue", "dogs"] },
     });
 
-    act(() => result.current.onClear());
+    act(() => result.current.onClear(multiSelectId));
     // selectedItems should be reset to empty
     expect(result.current.selectedItems).toEqual({});
   });