+++ src/hooks/useMultiSelect.ts
@@ -4,68 +4,64 @@ import {
   SelectedItems,
 } from "../components/MultiSelect/MultiSelect";
 
-interface useMultiSelectProps {
-  multiSelectId: string;
-  items: MultiSelectItem[];
-}
-
 /**
- * The useMultiSelect The hook returns an object containing all the props and state needed to handle the selectedItems.
- * That includes the functions onChange, onClear, onMixedStateChange for handling any changes to the selection of items and the current state of the selection: selectedItems.
+ * The useMultiSelect hook returns an object containing all the functions and state needed to handle the selectedItems of a `MultiSelect` component.
+ * It can be used in conjunction with one single `MultiSelect` component as well as a group of `MultiSelect`s in the `MultiSelectGroup` component.
+ * The returned object includes the functions onChange, onClear, onMixedStateChange for handling any changes to the selection of items
+ * and the current state of the selection: selectedItems.
  */
-export default function useMultiSelect({
-  multiSelectId,
-  items,
-}: useMultiSelectProps) {
+
+export default function useMultiSelect() {
   const [selectedItems, setSelectedItems] = useState<SelectedItems | {}>({});
 
-  function handleChange(itemId: string) {
+  /**
+   * handleChange is used for both MultiSelect variants. It handles the state for checkbox options that either have no child options or are child options themselves.
+   * It accepts two arguments - the id of the checkbox option and the id of the MultiSelect component.
+   */
+  const handleChange = (itemId: string, multiSelectId: string) => {
     let itemIds;
-    // Check if the id already exists in the state.
-    if (selectedItems[multiSelectId] !== undefined) {
-      let itemIdExists =
-        selectedItems[multiSelectId].items.indexOf(itemId) > -1;
-      // Make a copy of the existing items array.
+    // Check if the multiSelect already exists in the state.
+    if (selectedItems.hasOwnProperty(multiSelectId)) {
+      // Make a copy of the existing selected items.
       itemIds = selectedItems[multiSelectId].items.slice();
-      // If id exists, remove it from the array.
-      if (itemIdExists) {
+      // If itemId is in the selectedItems, remove it from the array.
+      if (selectedItems[multiSelectId].items.indexOf(itemId) > -1) {
         itemIds = itemIds.filter((id) => id !== itemId);
       } else {
         // Add it to the array, but modify the copy, not the original.
         itemIds.push(itemId);
       }
+      // If there were no items from the multiSelect in the selectedItems before
     } else {
+      // Add the itemId to the itemsIds
       itemIds = [];
       itemIds.push(itemId);
     }
-
+    // Update selectedItems on state to reflect the new selection
     setSelectedItems({
       ...selectedItems,
       [multiSelectId]: {
         items: itemIds,
       },
     });
-  }
+  };
 
   /**
-   * handleMixedStateChange is only used for the "dialog" variant. It handles the state for checkbox options with child options.
-   * It accepts one argument - the id of the parent checkbox option.
+   * handleMixedStateChange is used only for the "dialog" variant. It handles the state for checkbox options with child options.
+   * It accepts three arguments - the id of the (parent) checkbox option, the id of the MultiSelect component and the items array of the MultiSelect.
    */
-  function handleMixedStateChange(parentId: string) {
+  const handleMixedStateChange = (
+    parentId: string,
+    multiSelectId: string,
+    items: MultiSelectItem[]
+  ) => {
     // Build an array of child items.
-    let childItems = [];
-    items.forEach((item) => {
-      // If the parentId is an id within the all items
-      if (item.id === parentId) {
-        // Push each item into the childItems array
-        item.children.forEach((childItem: MultiSelectItem) => {
-          childItems.push(childItem.id);
-        });
-      }
-    });
+    const childItems = items
+      .filter((item: MultiSelectItem) => item.id === parentId)[0]
+      .children.map((child) => child.id);
 
     let newItems;
-    // If some items have already been selected
+    // If some items of the multiSelect are already selected
     if (selectedItems[multiSelectId] !== undefined) {
       // If all children of the parent are already selected
       if (
@@ -87,7 +83,7 @@ export default function useMultiSelect({
           ...selectedItems[multiSelectId].items,
         ];
       }
-      // If no items have been selected, the childItems will be the only ones added
+      // If no items of this multiSelect were selected before
     } else {
       newItems = childItems;
     }
@@ -98,11 +94,21 @@ export default function useMultiSelect({
         items: newItems,
       },
     });
-  }
+  };
 
-  function handleClear() {
-    setSelectedItems({});
-  }
+  /**
+   * handleClear is used for both MultiSelect variants. It will remove all selected items of specific MultiSelect component from the selectedItems array.
+   * It accepts one argument - the id of the MultiSelect component.
+   */
+  const handleClear = (multiSelectId: string) => {
+    let newSelectedItems = {};
+    for (let key of Object.keys(selectedItems)) {
+      if (key !== multiSelectId) {
+        newSelectedItems[key] = selectedItems[key];
+      }
+    }
+    setSelectedItems(newSelectedItems);
+  };
 
   return {
     selectedItems,